refactor: move core logic to `lib.rs` and expose necessary types
This commit restructures the project by moving the core logic and types from main.rs to lib.rs.
Changes include making `OutPutSummery`, `process_carte_version`, and `Index` public.
